Support for non-player economy accounts
Re-adds support for non-player economy accounts. After the transition to UUIDs, interacting with a shop sign owned by a non-player account would result in the transaction not going through. This code first checks if the player is a real player. If so, it continues as normal. If not, it sets the owner variable to a fake offline player of the account's name.